requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
APScheduler==3.11.0
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
click==8.1.7
discord.py==2.4.0
Django==5.1.3
django-environ==0.11.2
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.0
psycopg2-binary==2.9.10
sqlparse==0.5.1
tabulate==0.9.0
tzlocal==5.2
yarl==1.17.1
